refactor: deprecate feature-service-admin package (and bundle code in feature-service)
AFFECTS PACKAGES:
@esri/arcgis-rest-feature-service

BREAKING CHANGE:
deprecate feature-service-admin package (and bundle code in feature-service)
