--- packages/arcgis-rest-feature-service-admin/src/index.ts
@@ -1,2 +1,0 @@
-export * from "./create";
-export * from "./addTo";
