+++ packages/arcgis-rest-feature-service/package.json
@@ -16,12 +16,17 @@
     "tslib": "^1.9.3"
   },
   "devDependencies": {
+    "@esri/arcgis-rest-request": "^1.19.2",
     "@esri/arcgis-rest-common": "^1.19.2",
-    "@esri/arcgis-rest-request": "^1.19.2"
+    "@esri/arcgis-rest-auth": "^1.19.2",
+    "@esri/arcgis-rest-items": "^1.19.2"
+
   },
   "peerDependencies": {
+    "@esri/arcgis-rest-request": "^1.19.2",
     "@esri/arcgis-rest-common": "^1.19.2",
-    "@esri/arcgis-rest-request": "^1.19.2"
+    "@esri/arcgis-rest-auth": "^1.19.2",
+    "@esri/arcgis-rest-items": "^1.19.2"
   },
   "scripts": {
     "prepare": "npm run build",
