+++ packages/arcgis-rest-feature-service/src/admin/addToServiceDefinition.ts
