+++ packages/arcgis-rest-feature-service/src/admin/createServiceDefinition.ts
