+++ packages/arcgis-rest-feature-service/test/admin/createServiceDefinition.test.ts
@@ -3,10 +3,10 @@
 
 import * as fetchMock from "fetch-mock";
 
-import { createFeatureService } from "../src/create";
+import { createFeatureService } from "../../src/admin/createServiceDefinition";
 
-import { FeatureServiceResponse } from "./mocks/service";
-import { MoveToFolderResponse } from "./mocks/move";
+import { FeatureServiceResponse } from "../mocks/service";
+import { MoveToFolderResponse } from "../mocks/move";
 
 import { UserSession } from "@esri/arcgis-rest-auth";
 import { TOMORROW } from "@esri/arcgis-rest-auth/test/utils";
