+++ packages/arcgis-rest-feature-service/test/mocks/layerDefinition.ts
