+++ packages/arcgis-rest-feature-service/test/mocks/move.ts
